Fix attention_mask format for ChatGLM - convert boolean to integer tensor

+++ pipeline/dataset.py
@@ -87,7 +87,7 @@ class BackdoorData(Dataset):
         return dict(
             input_ids=all_input_ids,
             labels=all_labels,
-            attention_mask=all_input_ids.ne(tokenizer.pad_token_id),
+            attention_mask=all_input_ids.ne(tokenizer.pad_token_id).long(),
         )
 
     def __getitem__(self,index):
